Class RightPanel has been refactored
